Added shadows

+++ Apollo.Core/Display/AbstractClasses/AbstractColour.cs
@@ -2,10 +2,10 @@
 
 public abstract class AbstractColour
 {
-    public virtual float R { get; } = 0;
-    public virtual float G { get; } = 0;
-    public virtual float B { get; } = 0;
-    private const float Epsilon = 0.0001f;
+    public virtual double R { get; } = 0;
+    public virtual double G { get; } = 0;
+    public virtual double B { get; } = 0;
+    private const double Epsilon = 0.0001f;
 
     
     public override bool Equals(Object? obj)
@@ -61,7 +61,7 @@ public abstract class AbstractColour
         return new Colour(c1.R - c2.R, c1.G - c2.G, c1.B - c2.B);
     }
     
-    public static AbstractColour operator *(AbstractColour c1, float scalar)
+    public static AbstractColour operator *(AbstractColour c1, double scalar)
     {
         return new Colour(c1.R * scalar, c1.G * scalar, c1.B * scalar);
     }
