+++ Apollo.Core/Display/ColourPresets/Black.cs
@@ -4,7 +4,7 @@ namespace Apollo.Display.ColourPresets;
 
 public class Black : AbstractColour
 {
-    public override float R { get; } = 0;
-    public override float G { get; } = 0;
-    public override float B { get; } = 0;
+    public override double R { get; } = 0;
+    public override double G { get; } = 0;
+    public override double B { get; } = 0;
 }
