+++ Apollo.Core/Display/ColourPresets/Red.cs
@@ -4,7 +4,7 @@ namespace Apollo.Display.ColourPresets;
 
 public class Red : AbstractColour
 {
-    public override float R { get; } = 1;
-    public override float G { get; } = 0;
-    public override float B { get; } = 0;
+    public override double R { get; } = 1;
+    public override double G { get; } = 0;
+    public override double B { get; } = 0;
 }
