+++ Apollo.Core/Display/ColourPresets/White.cs
@@ -4,7 +4,7 @@ namespace Apollo.Display.ColourPresets;
 
 public class White : AbstractColour
 {
-    public override float R { get; } = 1;
-    public override float G { get; } = 1;
-    public override float B { get; } = 1;
+    public override double R { get; } = 1;
+    public override double G { get; } = 1;
+    public override double B { get; } = 1;
 }
