+++ Apollo.Core/Lighting/Material.cs
@@ -7,12 +7,12 @@ namespace Apollo.Lighting;
 public class Material
 {
     public AbstractColour Colour { get; set; }
-    public float Ambient { get; set; }
-    public float Diffuse { get; set; }
-    public float Specular { get; set; }
-    public float Shininess { get; set; }
+    public double Ambient { get; set; }
+    public double Diffuse { get; set; }
+    public double Specular { get; set; }
+    public double Shininess { get; set; }
 
-    public Material(AbstractColour colour, float ambient, float diffuse, float specular, float shininess)
+    public Material(AbstractColour colour, double ambient, double diffuse, double specular, double shininess)
     {
         (Colour, Ambient, Diffuse, Specular, Shininess) = (colour, ambient, diffuse, specular, shininess);
     }
